+++ test/clojure_lsp/features/rename_test.clj
@@ -3,7 +3,8 @@
    [clojure-lsp.db :as db]
    [clojure-lsp.feature.rename :as f.rename]
    [clojure-lsp.test-helper :as h]
-   [clojure.test :refer [deftest is testing]]))
+   [clojure.test :refer [deftest is testing]]
+   [medley.core :as medley]))
 
 (h/reset-db-after-test)
 
@@ -81,3 +82,45 @@
         (is (= {(h/file-uri "file:///a.cljc") [{:new-text ":a/g" :range (h/->range a-b-start a-b-stop)}
                                                {:new-text ":g" :range (h/->range b-ns-start b-ns-stop)}]}
                changes))))))
+
+(deftest rename-namespaces
+  (testing "when client has valid source-paths but no document-changes capability"
+    (h/clean-db!)
+    (swap! db/db medley/deep-merge
+           {:project-root-uri (h/file-uri "file:///my-project")
+            :settings {:source-paths #{(h/file-path "/my-project/src") (h/file-path "/my-project/test")}}
+            :client-capabilities {:workspace {:workspace-edit {:document-changes false}}}})
+    (h/load-code-and-locs "(ns foo.bar-baz)" (h/file-uri "file:///my-project/src/foo/bar_baz.clj"))
+    (h/assert-submap
+      {:error {:code :invalid-params
+               :message "Can't rename namespace, client does not support file renames."}}
+      (f.rename/rename (h/file-uri "file:///my-project/src/foo/bar_baz.clj") "foo.baz-qux" 1 5 db/db)))
+  (testing "when client has document-changes capability but no valid source-paths"
+    (h/clean-db!)
+    (swap! db/db medley/deep-merge
+           {:project-root-uri (h/file-uri "file:///my-project")
+            :settings {:source-paths #{(h/file-path "/my-project/bla")}}
+            :client-capabilities {:workspace {:workspace-edit {:document-changes true}}}})
+    (h/load-code-and-locs "(ns foo.bar-baz)" (h/file-uri "file:///my-project/src/foo/bar_baz.clj"))
+    (h/assert-submap
+      {:error {:code :invalid-params
+               :message "Can't rename namespace, invalid source-paths. Are project :source-paths configured correctly?"}}
+      (f.rename/rename (h/file-uri "file:///my-project/src/foo/bar_baz.clj") "foo.baz-qux" 1 5 db/db)))
+  (testing "when source-paths are valid and client capabilities has document-changes"
+    (h/clean-db!)
+    (swap! db/db medley/deep-merge
+           {:project-root-uri (h/file-uri "file:///my-project")
+            :settings {:source-paths #{(h/file-path "/my-project/src") (h/file-path "/my-project/test")}}
+            :client-capabilities {:workspace {:workspace-edit {:document-changes true}}}})
+    (h/load-code-and-locs "(ns foo.bar-baz)" (h/file-uri "file:///my-project/src/foo/bar_baz.clj"))
+    (is (= {:document-changes
+            [{:text-document {:version 0
+                              :uri (h/file-uri "file:///my-project/src/foo/bar_baz.clj")}
+              :edits [{:range
+                       {:start {:line 0 :character 4}
+                        :end {:line 0 :character 15}}
+                       :new-text "foo.baz-qux"}]}
+             {:kind "rename"
+              :old-uri (h/file-uri "file:///my-project/src/foo/bar_baz.clj")
+              :new-uri (h/file-uri "file:///my-project/src/foo/baz_qux.clj")}]}
+           (f.rename/rename (h/file-uri "file:///my-project/src/foo/bar_baz.clj") "foo.baz-qux" 1 5 db/db)))))